Update SMILE to VC 2009 Add more flexible multithreading to BNCreator Fix a few Windows and GCC 4.3 build problems Add a draft MedianMultiples implementation   Funkiness!  Use at your own risk Add subsampling to MIer Add more flexible PCL normalization toe Normalizer

+++ src/coalescemotifs.h
@@ -157,7 +157,7 @@ public:
 
 		if( iOne == iTwo )
 			return ( fAllowDuplicates ? iOne : -1 );
-		priiMerged.first = min( iOne, iTwo );
+		priiMerged.first = (uint32_t)min( iOne, iTwo );
 		priiMerged.second = max( iOne, iTwo );
 		if( ( iterMerged = m_mappriiiMerged.find( priiMerged ) ) != m_mappriiiMerged.end( ) )
 			return ( fAllowDuplicates ? iterMerged->second : -1 );
